requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
asyncio-throttle==1.0.2
requests==2.31.0
python-dotenv==1.0.0
